Add accordion control to categorize templates In this change, we have add accordion control which can be used to organize template snippets by categories.
One thing need to take caution is that, our product is using anchor(href) for navigation. Since anchor does not support nesting, so some anchor based widgets in our templates, like tab control or accordion could not work.
As an alternative, we have used javascript to simulate the corresponding behavior.